apps/phase1-openai-agents/pyproject.toml
[project]
name = "phase1-openai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.13.2",
    "python-dotenv>=1.2.2",
    "shared-prompts",
    "tavily-python>=0.7.23",
]

[tool.uv.sources]
shared-prompts = { path = "../../packages/shared-prompts" }
